+++ console-ui/src/pages/AuthorityControl/PermissionsManagement/PermissionsManagement.js
@@ -16,7 +16,16 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Button, Dialog, Pagination, Table, ConfigProvider } from '@alifd/next';
+import {
+  Button,
+  Dialog,
+  Pagination,
+  Table,
+  ConfigProvider,
+  Form,
+  Input,
+  Switch,
+} from '@alifd/next';
 import { connect } from 'react-redux';
 import { getPermissions, createPermission, deletePermission } from '../../../reducers/authority';
 import { getNamespaces } from '../../../reducers/namespace';
@@ -51,7 +60,10 @@ class PermissionsManagement extends React.Component {
       pageNo: 1,
       pageSize: 9,
       createPermission: false,
+      defaultFuzzySearch: true,
+      role: '',
     };
+    this.handleDefaultFuzzySwitchChange = this.handleDefaultFuzzySwitchChange.bind(this);
   }
 
   componentDidMount() {
@@ -62,8 +74,18 @@ class PermissionsManagement extends React.Component {
   getPermissions() {
     this.setState({ loading: true });
     const { pageNo, pageSize } = this.state;
+    let role = this.state.role;
+    let search = 'accurate';
+    if (this.state.defaultFuzzySearch) {
+      if (role && role !== '') {
+        role = '*' + role + '*';
+      }
+    }
+    if (role && role.indexOf('*') !== -1) {
+      search = 'blur';
+    }
     this.props
-      .getPermissions({ pageNo, pageSize })
+      .getPermissions({ pageNo, pageSize, role, search })
       .then(() => {
         if (this.state.loading) {
           this.setState({ loading: false });
@@ -85,24 +107,57 @@ class PermissionsManagement extends React.Component {
     }[action];
   }
 
+  handleDefaultFuzzySwitchChange() {
+    this.setState({ defaultFuzzySearch: !this.state.defaultFuzzySearch });
+  }
+
   render() {
     const { permissions, namespaces = [], locale } = this.props;
     const { loading, pageSize, pageNo, createPermissionVisible } = this.state;
     return (
       <>
         <RegionGroup left={locale.privilegeManagement} />
-        <div className="filter-panel">
-          <Button
-            type="primary"
-            onClick={() => this.setState({ createPermissionVisible: true })}
-            style={{ marginRight: 20 }}
-          >
-            {locale.addPermission}
-          </Button>
-          <Button type="secondary" onClick={() => this.getPermissions()}>
-            {locale.refresh}
-          </Button>
-        </div>
+        <Form inline>
+          <Form.Item label="角色名">
+            <Input
+              value={this.state.role}
+              htmlType="text"
+              placeholder={this.state.defaultFuzzySearch ? locale.defaultFuzzyd : locale.fuzzyd}
+              style={{ width: 200 }}
+              onChange={role => {
+                this.setState({ role });
+              }}
+            />
+          </Form.Item>
+          <Form.Item label="默认模糊匹配">
+            <Switch
+              checkedChildren=""
+              unCheckedChildren=""
+              defaultChecked={this.state.defaultFuzzySearch}
+              onChange={this.handleDefaultFuzzySwitchChange}
+              title={'自动在搜索参数前后加上*'}
+            />
+          </Form.Item>
+          <Form.Item label={''}>
+            <Button
+              type={'primary'}
+              style={{ marginRight: 10 }}
+              onClick={() => this.getPermissions()}
+              data-spm-click={'gostr=/aliyun;locaid=dashsearch'}
+            >
+              {locale.query}
+            </Button>
+          </Form.Item>
+          <Form.Item style={{ float: 'right' }}>
+            <Button
+              type="primary"
+              onClick={() => this.setState({ createPermissionVisible: true })}
+              style={{ marginRight: 20 }}
+            >
+              {locale.addPermission}
+            </Button>
+          </Form.Item>
+        </Form>
         <Table dataSource={permissions.pageItems} loading={loading} maxBodyHeight={476} fixedHeader>
           <Table.Column title={locale.role} dataIndex="role" />
           <Table.Column