+++ naming/src/main/java/com/alibaba/nacos/naming/selector/NoneSelector.java
@@ -24,7 +24,7 @@ import com.alibaba.nacos.api.selector.Selector;
 import java.util.List;
 
 /**
- * Selector with no filtering. The logic is same to {@link com.alibaba.nacos.naming.selector.v1.NoneSelector}.
+ * Selector with no filtering.
  *
  * @author chenglu
  * @date 2021-08-04 13:28